fix(contract): run fixture per-job and fail loud when required
- Restructure workflow: each language job starts its own fixture on the
  same runner and tears it down; the old shared 'fixture-ready' job ran on
  a different runner, so Rust/JS/Flutter contract tests silently skipped.
- Runners now FAIL if ECHOMIRROR_CONTRACT_SPEC is set and the fixture is
  unreachable (CI = required, never a green skip). Result: the fixture-wait
  step also hard-fails instead of letting a skip through.
- Swift job no longer touches docker (its contract tests are FFI-local).
- Fix Swift ContractTests type error: MoodEntry.score is UInt8, not Int.

+++ packages/swift/EchoMirrorSDK/Tests/EchoMirrorSDKTests/ContractTests.swift
@@ -53,10 +53,10 @@ final class ContractTests: XCTestCase {
 
     func testMoodScoreBoundsFromContract() throws {
         let spec = try Self.loadSpec()
-        let score = spec.logMoodBody["score"] as? Int ?? 8
+        let score = UInt8(spec.logMoodBody["score"] as? Int ?? 8)
         let mood = try MoodClient(config: EchoMirrorConfig(apiKey: "contract-test-key", network: .testnet))
 
-        XCTAssertTrue(mood.isValidScore(UInt8(score)))
+        XCTAssertTrue(mood.isValidScore(score))
         XCTAssertTrue(mood.isValidScore(1))
         XCTAssertTrue(mood.isValidScore(10))
         XCTAssertFalse(mood.isValidScore(0))
@@ -66,14 +66,14 @@ final class ContractTests: XCTestCase {
     func testMoodLogBridgesFromContract() async throws {
         let spec = try Self.loadSpec()
         let body = spec.logMoodBody
-        let score = body["score"] as? Int ?? 8
+        let score = UInt8(body["score"] as? Int ?? 8)
         let note = body["note"] as? String ?? "Great day"
         let tags = body["tags"] as? [String] ?? ["work", "proud"]
 
         let mood = try MoodClient(config: EchoMirrorConfig(apiKey: "contract-test-key", network: .testnet))
         let entry = try await mood.logMood(
             userId: spec.moodUserId,
-            score: UInt8(score),
+            score: score,
             note: note,
             tags: tags
         )
